# backend/src/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Under the Sea",
    "description": "싸피 서울 6기 특화 프로젝트 - UTS",
    "version": "1.0.0"
  },
  "host": "uts_url:8080/api",
  "servers": [
    {
      "url": "http://uts_url:8080"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/market/artists": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "sortby",
            "in": "query",
            "type": "string"
          },
          {
            "name": "category",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/market/nfts": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "sortby",
            "in": "query",
            "type": "string"
          },
          {
            "name": "category",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/nft/minting": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userSeq": {
                  "example": "any"
                },
                "editionName": {
                  "example": "any"
                },
                "editionImage": {
                  "example": "any"
                },
                "editionDescription": {
                  "example": "any"
                },
                "editionRoyalty": {
                  "example": "any"
                },
                "editionTotal": {
                  "example": "any"
                },
                "salePrice": {
                  "example": "any"
                },
                "nftId": {
                  "example": "any"
                },
                "nftTransactionId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/nft/info": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "nftSeq",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/admin/search/user": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userWalletAddress",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/artist/accept": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userSeq": {
                  "example": "any"
                },
                "commonCodeSeq": {
                  "example": "any"
                },
                "artistDescription": {
                  "example": "any"
                },
                "artistSns": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/artist/cancel": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userSeq": {
                  "example": "any"
                },
                "artistSeq": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/admin/report/cancel": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "reportSeq": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}